Isolate mina and jsch ssh test backend in own classes
