+++ java/com/google/gerrit/acceptance/SshSession.java
@@ -14,140 +14,32 @@
 
 package com.google.gerrit.acceptance;
 
-import static com.google.common.base.Preconditions.checkState;
-import static com.google.common.io.RecursiveDeleteOption.ALLOW_INSECURE;
 import static com.google.common.truth.Truth.assertThat;
 import static com.google.common.truth.Truth.assertWithMessage;
-import static com.google.gerrit.server.config.SshClientImplementation.JSCH;
-import static com.google.gerrit.server.config.SshClientImplementation.getFromEnvironment;
-import static java.nio.charset.StandardCharsets.UTF_8;
 
-import com.google.common.io.CharSink;
-import com.google.common.io.Files;
-import com.google.common.io.MoreFiles;
 import com.google.gerrit.acceptance.testsuite.account.TestAccount;
 import com.google.gerrit.acceptance.testsuite.account.TestSshKeys;
-import com.google.gerrit.server.config.SshClientImplementation;
-import com.jcraft.jsch.ChannelExec;
-import com.jcraft.jsch.JSch;
-import com.jcraft.jsch.Session;
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
 import java.net.InetSocketAddress;
-import java.security.KeyPair;
-import java.util.Arrays;
-import java.util.Scanner;
-import org.apache.sshd.common.config.keys.writer.openssh.OpenSSHKeyPairResourceWriter;
-import org.eclipse.jgit.transport.URIish;
-import org.eclipse.jgit.transport.sshd.DefaultProxyDataFactory;
-import org.eclipse.jgit.transport.sshd.JGitKeyCache;
-import org.eclipse.jgit.transport.sshd.SshdSession;
-import org.eclipse.jgit.transport.sshd.SshdSessionFactory;
-import org.eclipse.jgit.util.FS;
 
-public class SshSession {
-  private static final int TIMEOUT = 100000;
-
-  private final SshClientImplementation client;
-  private final TestSshKeys sshKeys;
-  private final InetSocketAddress addr;
-  private final TestAccount account;
-  private SshdSession sshdSession;
-  private Session jschSession;
-  private File userhome;
-  private String error;
+public abstract class SshSession {
+  protected final TestSshKeys sshKeys;
+  protected final InetSocketAddress addr;
+  protected final TestAccount account;
+  protected String error;
 
   public SshSession(TestSshKeys sshKeys, InetSocketAddress addr, TestAccount account) {
-    this.client = getFromEnvironment();
     this.sshKeys = sshKeys;
     this.addr = addr;
     this.account = account;
   }
 
-  public void open() throws Exception {
-    if (client == JSCH) {
-      getJschSession();
-    } else {
-      getMinaSession();
-    }
-  }
+  public abstract void open() throws Exception;
 
-  public String exec(String command) throws Exception {
-    return client == JSCH ? execJsch(command) : execMina(command);
-  }
+  public abstract void close();
 
-  public int execAndReturnStatus(String command) throws Exception {
-    return client == JSCH ? execJschAndReturnStatus(command) : execMinaAndReturnStatus(command);
-  }
+  public abstract String exec(String command) throws Exception;
 
-  @SuppressWarnings("resource")
-  private String execMina(String command) throws Exception {
-    Process process = getMinaSession().exec(command, TIMEOUT);
-    InputStream in = process.getInputStream();
-    InputStream err = process.getErrorStream();
-
-    Scanner s = new Scanner(err, UTF_8.name()).useDelimiter("\\A");
-    error = s.hasNext() ? s.next() : null;
-
-    s = new Scanner(in, UTF_8.name()).useDelimiter("\\A");
-    return s.hasNext() ? s.next() : "";
-  }
-
-  @SuppressWarnings("resource")
-  private String execJsch(String command) throws Exception {
-    ChannelExec channel = (ChannelExec) getJschSession().openChannel("exec");
-    try {
-      channel.setCommand(command);
-      InputStream in = channel.getInputStream();
-      InputStream err = channel.getErrStream();
-      channel.connect();
-
-      Scanner s = new Scanner(err, UTF_8.name()).useDelimiter("\\A");
-      error = s.hasNext() ? s.next() : null;
-
-      s = new Scanner(in, UTF_8.name()).useDelimiter("\\A");
-      return s.hasNext() ? s.next() : "";
-    } finally {
-      channel.disconnect();
-    }
-  }
-
-  @SuppressWarnings("resource")
-  public int execMinaAndReturnStatus(String command) throws Exception {
-    Process process = getMinaSession().exec(command, 0);
-    InputStream in = process.getInputStream();
-    InputStream err = process.getErrorStream();
-
-    Scanner s = new Scanner(err, UTF_8.name()).useDelimiter("\\A");
-    error = s.hasNext() ? s.next() : null;
-
-    s = new Scanner(in, UTF_8.name()).useDelimiter("\\A");
-    try {
-      return process.exitValue();
-    } catch (IllegalThreadStateException e) {
-      // SSH command was interrupted
-      return -1;
-    }
-  }
-
-  @SuppressWarnings("resource")
-  public int execJschAndReturnStatus(String command) throws Exception {
-    ChannelExec channel = (ChannelExec) getJschSession().openChannel("exec");
-    try {
-      channel.setCommand(command);
-      InputStream err = channel.getErrStream();
-      channel.connect();
-
-      Scanner s = new Scanner(err, UTF_8.name()).useDelimiter("\\A");
-      error = s.hasNext() ? s.next() : null;
-      return channel.getExitStatus();
-    } finally {
-      channel.disconnect();
-    }
-  }
+  public abstract int execAndReturnStatus(String command) throws Exception;
 
   private boolean hasError() {
     return error != null;
@@ -170,85 +62,7 @@ public class SshSession {
     assertThat(getError()).contains(error);
   }
 
-  public void close() {
-    if (client == JSCH) {
-      if (jschSession != null) {
-        jschSession.disconnect();
-        jschSession = null;
-      }
-    } else {
-      if (sshdSession != null) {
-        sshdSession.disconnect();
-        sshdSession = null;
-      }
-    }
-  }
-
-  private Session getJschSession() throws Exception {
-    if (jschSession == null) {
-      KeyPair keyPair = sshKeys.getKeyPair(account);
-      JSch jsch = new JSch();
-      jsch.addIdentity(
-          "KeyPair", TestSshKeys.privateKey(keyPair), TestSshKeys.publicKeyBlob(keyPair), null);
-      String username = getUsername();
-      jschSession = jsch.getSession(username, addr.getAddress().getHostAddress(), addr.getPort());
-      jschSession.setConfig("StrictHostKeyChecking", "no");
-      jschSession.connect();
-    }
-    return jschSession;
-  }
-
-  private SshdSession getMinaSession() throws Exception {
-    if (sshdSession == null) {
-      String username = getUsername();
-
-      URIish uri =
-          new URIish(
-              "ssh://"
-                  + username
-                  + "@"
-                  + addr.getAddress().getHostAddress()
-                  + ":"
-                  + addr.getPort());
-
-      // TODO(davido): Switch to memory only key resolving mode.
-      userhome = Files.createTempDir();
-
-      FS fs = FS.DETECTED.setUserHome(userhome);
-      File sshDir = new File(userhome, ".ssh");
-      sshDir.mkdir();
-      OpenSSHKeyPairResourceWriter keyPairWriter = new OpenSSHKeyPairResourceWriter();
-      try (OutputStream out = new FileOutputStream(new File(sshDir, "id_ecdsa"))) {
-        keyPairWriter.writePrivateKey(sshKeys.getKeyPair(account), null, null, out);
-      }
-
-      // TODO(davido): Disable programmatically host key checking: "StrictHostKeyChecking: no" mode.
-      CharSink configFile = Files.asCharSink(new File(sshDir, "config"), UTF_8);
-      configFile.writeLines(Arrays.asList("Host *", "StrictHostKeyChecking no"));
-
-      JGitKeyCache keyCache = new JGitKeyCache();
-      try (SshdSessionFactory factory =
-          new SshdSessionFactory(keyCache, new DefaultProxyDataFactory())) {
-        factory.setHomeDirectory(userhome);
-        factory.setSshDirectory(sshDir);
-
-        sshdSession = factory.getSession(uri, null, fs, TIMEOUT);
-
-        sshdSession.addCloseListener(
-            future -> {
-              try {
-                MoreFiles.deleteRecursively(userhome.toPath(), ALLOW_INSECURE);
-              } catch (IOException e) {
-                e.printStackTrace();
-              }
-            });
-      }
-    }
-    return sshdSession;
-  }
-
   public String getUrl() {
-    checkState(sshdSession != null || jschSession != null, "session must be opened");
     StringBuilder b = new StringBuilder();
     b.append("ssh://");
     b.append(account.username().get());
@@ -259,22 +73,12 @@ public class SshSession {
     return b.toString();
   }
 
-  public TestAccount getAccount() {
-    return account;
-  }
-
-  public File getUserhome() {
-    return userhome;
-  }
-
-  private String getUsername() {
-    String username =
-        account
-            .username()
-            .orElseThrow(
-                () ->
-                    new IllegalStateException(
-                        "account " + account.accountId() + " must have a username to use SSH"));
-    return username;
+  protected String getUsername() {
+    return account
+        .username()
+        .orElseThrow(
+            () ->
+                new IllegalStateException(
+                    "account " + account.accountId() + " must have a username to use SSH"));
   }
 }
